update plugin install process
