+++ scripts/android/before_compile.js
