+++ scripts/package.json
@@ -4,7 +4,10 @@
   "description": "Just used for cordova scripts",
   "main": "",
   "dependencies": {
-    "fs-extra-promise": "^0.2.0"
+    "child-process-promise": "^1.1.0",
+    "fs-extra-promise": "^0.2.0",
+    "lie": "^3.1.0",
+    "xcode": "^0.8.9"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
