+++ thali/install/package.json
@@ -9,8 +9,7 @@
     "jxc": "^1.0.14",
     "lie": "^3.1.0",
     "request": "^2.74.0",
-    "unzip": "^0.1.11",
-    "xcode": "^0.8.9"
+    "unzip": "^0.1.11"
   },
   "scripts": {
     "setupUnit": "./setUpTests.sh UnitTest_app.js",
